arreglar id3 en ubuntu
